# Fix arc handling for `VMobjectFromSVGPath` when a matrix transform is present in the SVG