@jkjkil4 2025-03-17
By the way, this solution reduces the number of points needed to describe an arc. If anything depends on the original point count, it might break, though I'm not sure exactly what relies on it.